Issue 2 Improvements 1 HTML nothing python: refactor new part in other python file and call that module in main (modularity)
